Stop trying to read if no inputs are connected
If no inputs are connected, the `_read` function still keeps looping ( using `setImmediate(this._read.bind(this));` ) to wait for more samples to be available. This can cause the CPU usage to go up (60% on my 2014 MBP) even when no audio is being played.

This fix stops the looping if `this.inputs` array is empty and starts it again when new inputs are connected.